requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-jwt
djangorestframework-filters
django_filter